requirements.txt
python-telegram-bot
yt-dlp>=2026.3.17